WIP to allow more concise commits for refactor
