+++ MobileSyncExplorerKotlinTemplate/app/src/main/java/com/salesforce/mobilesyncexplorerkotlintemplate/contacts/detailscomponent/ContactDetailsUiState.kt
@@ -5,7 +5,6 @@ import com.salesforce.mobilesyncexplorerkotlintemplate.core.ui.state.SObjectUiSy
 import com.salesforce.mobilesyncexplorerkotlintemplate.model.contacts.ContactObject
 
 sealed interface ContactDetailsUiState {
-    val curDialogUiState: DialogUiState?
     val dataOperationIsActive: Boolean // TODO it may be a good idea to break this up into discrete flags for the different types of data operations, and then the UI just shows the spinner if any of the flags are true.
     val doingInitialLoad: Boolean
 
@@ -19,7 +18,6 @@ sealed interface ContactDetailsUiState {
         val isEditingEnabled: Boolean,
         val shouldScrollToErrorField: Boolean,
 
-        override val curDialogUiState: DialogUiState?,
         override val dataOperationIsActive: Boolean,
         override val doingInitialLoad: Boolean = false
     ) : ContactDetailsUiState {
@@ -30,24 +28,20 @@ sealed interface ContactDetailsUiState {
     }
 
     data class NoContactSelected(
-        override val curDialogUiState: DialogUiState?,
         override val dataOperationIsActive: Boolean,
         override val doingInitialLoad: Boolean = false
     ) : ContactDetailsUiState
 }
 
 fun ContactDetailsUiState.copy(
-    curDialogUiState: DialogUiState? = this.curDialogUiState,
     dataOperationIsActive: Boolean = this.dataOperationIsActive,
     doingInitialLoad: Boolean = this.doingInitialLoad,
 ) = when (this) {
     is ContactDetailsUiState.NoContactSelected -> this.copy(
-        curDialogUiState = curDialogUiState,
         dataOperationIsActive = dataOperationIsActive,
         doingInitialLoad = doingInitialLoad,
     )
     is ContactDetailsUiState.ViewingContactDetails -> this.copy(
-        curDialogUiState = curDialogUiState,
         dataOperationIsActive = dataOperationIsActive,
         doingInitialLoad = doingInitialLoad,
     )
