+++ MobileSyncExplorerKotlinTemplate/app/src/main/java/com/salesforce/mobilesyncexplorerkotlintemplate/core/repos/usecases/UndeleteUseCase.kt
@@ -10,26 +10,17 @@ import kotlinx.coroutines.flow.flow
 sealed interface UndeleteResponse<T : SObject> {
     class Started<T : SObject> : UndeleteResponse<T>
     data class UndeleteSuccess<T : SObject>(val record: SObjectRecord<T>) : UndeleteResponse<T>
-    data class Finished<T : SObject>(val exception: RepoOperationException? = null) :
-        UndeleteResponse<T>
 }
 
 class UndeleteUseCase<T : SObject>(private val repo: SObjectSyncableRepo<T>) {
+    @Throws(RepoOperationException::class)
     operator fun invoke(id: String): Flow<UndeleteResponse<T>> = runUndelete(id = id)
 
+    @Throws(RepoOperationException::class)
     private fun runUndelete(id: String): Flow<UndeleteResponse<T>> = flow {
-        var exception: RepoOperationException? = null
-        try {
-            emit(UndeleteResponse.Started())
+        emit(UndeleteResponse.Started())
+        val record = repo.locallyUndelete(id = id)
+        emit(UndeleteResponse.UndeleteSuccess(record))
 
-            val record = repo.locallyUndelete(id = id)
-
-            emit(UndeleteResponse.UndeleteSuccess(record))
-
-        } catch (ex: RepoOperationException) {
-            exception = ex
-        } finally {
-            emit(UndeleteResponse.Finished(exception = exception))
-        }
     }
 }
